Bug 569087: Make the EPP common feature a root feature
This allows it to be uninstalled, which allows the workaround
for JAXB problems to be uninstalled if there are unintended consequences
of that.

Includes updating the label to expand EPP to Eclipse Packaging Project.

This has the side effect that Oomph can be uninstalled as it is a child
feature of the common feature. Making Oomph a root feature is probably
something worth considering separately too.

Change-Id: I96a2e3f30771734cb51bac4802a0d8f3362fff87