[releng] Update to Tycho 2.7.1 for products with pgp-signed content
Use the repository as the source when materializing a product, rather
than the targetPlatform to ensure that the product's artifacts.xml
contains the pgp.publicKeys property for pgp-signed artifacts.

For this to work, the repository for the product must be created and it
must include all the requirements needed by the product.

Change-Id: Id074da4d70ffc4b4f88a6e03861d07acd7fa4711
Signed-off-by: Ed Merks <ed.merks@gmail.com>