Don't add slf4j.simple to products that don't include slf4j already
This is a follow up to #27 to workaround the committers, php and scout
products not including slf4j, so therefore not being able to have
start levels for them.

The org.apache.aries.spifly.dynamic.bundle start level for these bundles is
kept to allow installing, for example m2e logback.