Update to Tycho 4.0.8-SNAPSHOT to process update.id advice
All the update.id advice was commented because Tycho 4.0.7 throws
exception; this re-enables all those.
Also, each p2.inf for each product is renamed to epp.p2.inf which avoids
a related bug parsing the advice file for the project as a whole as
opposed to parse the product-specific advice.
Define a file association in the setup so it's easy to open an
eep.p2.inf with the properties editor.