Update check incubation step
I have not been doing the check incubation step for a while because
everything contains incubating components. This doesn't seem likely
to change anytime soon, so I have updated RELEASING.md to reflect
current state.

However, during a cross check I found that modeling did include
the required incubating components warning in the feature.properties
but was missing it in the product aboutText blurb. This commit
remedies that too.