Persist default value for DRY_RUN when promoting build
The decision was to make dry run off by default as it is primarily
useful when testing changes to the scripts which isn't common.

The change was previously made in the [job
config](https://ci.eclipse.org/packaging/job/promote-a-build/configure)
on Jenkins directly, but it will be overridden on the next build
because the Jenkinsfile controls that setting.